+++ SplunkAgent/Sources/SplunkAgent/Public API/API-1.0-Session.swift
@@ -54,6 +54,7 @@ extension Session {
         let user = owner.currentUser
 
         let anonymousUserId: String?
+
         if user.trackingMode != .noTracking {
             anonymousUserId = user.userIdentifier
         }