+++ SplunkCommon/Sources/SplunkCommon/Events/ActivityTracker.swift
@@ -1,6 +1,7 @@
 //
+//
 /*
-Copyright 2025 Splunk Inc.
+Copyright 2026 Splunk Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -15,9 +16,13 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-/// Defines list of attributes that change continuously over time.
-public protocol RuntimeAttributes: AnyObject {
+import Foundation
 
-    /// A list of attributes to use at signal start.
-    var all: [String: Any] { get }
+/// A protocol for tracking activity within a session.
+public protocol ActivityTracker {
+
+    /// Records a telemetry activity event for the current session.
+    ///
+    /// - Parameter date: The timestamp of the activity.
+    func trackActivity(at date: Date)
 }