+++ SplunkOpenTelemetry/Sources/SplunkOpenTelemetry/Log Processors/OTLPAttributesLogRecordProcessor.swift
@@ -25,6 +25,7 @@ public class OTLPAttributesLogRecordProcessor: LogRecordProcessor {
 
     // MARK: - Private
 
+    private let activityTracker: ActivityTracker?
     private let proxy: LogRecordProcessor
     private unowned let runtimeAttributes: any RuntimeAttributes
 
@@ -39,16 +40,17 @@ public class OTLPAttributesLogRecordProcessor: LogRecordProcessor {
     ///
     /// - Note: The processor itself does not own the object with runtime attributes.
     ///         So, ensuring its existence outside this processor is always necessary.
-    public init(proxy: LogRecordProcessor, with runtimeAttributes: RuntimeAttributes) {
+    public init(proxy: LogRecordProcessor, with runtimeAttributes: RuntimeAttributes, activityTracker: ActivityTracker) {
         self.proxy = proxy
         self.runtimeAttributes = runtimeAttributes
+        self.activityTracker = activityTracker
     }
 
 
     // MARK: - LogRecordProcessor methods
 
     public func onEmit(logRecord: OpenTelemetrySdk.ReadableLogRecord) {
-        runtimeAttributes.trackActivity()
+        activityTracker?.trackActivity(at: Date())
 
         var updatedAttributes = logRecord.attributes